+++ modules/client/jvm/src/main/scala/com/azavea/stac4s/api/client/SearchFilters.scala
@@ -8,6 +8,8 @@ import eu.timepit.refined.types.numeric.NonNegInt
 import geotrellis.vector.{io => _, _}
 import io.circe._
 import io.circe.refined._
+import monocle.Lens
+import monocle.macros.GenLens
 
 case class SearchFilters(
     bbox: Option[Bbox] = None,
@@ -21,6 +23,7 @@ case class SearchFilters(
 )
 
 object SearchFilters extends ClientCodecs {
+  implicit val paginationTokenLens: Lens[SearchFilters, Option[PaginationToken]] = GenLens[SearchFilters](_.next)
 
   implicit val searchFiltersDecoder: Decoder[SearchFilters] = { c =>
     for {