+++ modules/client/jvm/src/main/scala/com/azavea/stac4s/api/client/StacClient.scala
@@ -1,3 +1,0 @@
-package com.azavea.stac4s.api.client
-
-trait StacClient[F[_]] extends StacClientF[F, SearchFilters]