+++ modules/client/jvm/src/main/scala/com/azavea/stac4s/api/client/client.scala
@@ -1,5 +1,0 @@
-package com.azavea.stac4s.api
-
-package object client {
-  type SttpStacClient[F[_]] = SttpStacClientF[F, SearchFilters]
-}