+++ modules/client/shared/src/main/scala/com/azavea/stac4s/api/client/StacClientF.scala
@@ -5,12 +5,15 @@ import com.azavea.stac4s._
 import eu.timepit.refined.types.string.NonEmptyString
 
 trait StacClientF[F[_], S] {
-  def search: F[List[StacItem]]
-  def search(filter: S): F[List[StacItem]]
-  def collections: F[List[StacCollection]]
   def collection(collectionId: NonEmptyString): F[StacCollection]
-  def items(collectionId: NonEmptyString): F[List[StacItem]]
   def item(collectionId: NonEmptyString, itemId: NonEmptyString): F[StacItem]
   def itemCreate(collectionId: NonEmptyString, item: StacItem): F[StacItem]
   def collectionCreate(collection: StacCollection): F[StacCollection]
+}
+
+trait StreamingStacClientF[F[_], G[_], S] extends StacClientF[F, S] {
+  def search: G[StacItem]
+  def search(filter: S): G[StacItem]
+  def collections: G[StacCollection]
+  def items(collectionId: NonEmptyString): G[StacItem]
 }