+++ project/Versions.scala
@@ -2,10 +2,19 @@ import sbt._
 import sbt.Keys._
 
 object Versions {
-  val Cats                    = "2.6.0"
+
+  private def ver(for212: String, for213: String) = Def.setting {
+    CrossVersion.partialVersion(scalaVersion.value) match {
+      case Some((2, 12)) => for212
+      case Some((2, 13)) => for213
+      case _             => sys.error("not good")
+    }
+  }
+
+  val Cats                    = "2.6.1"
   val Circe                   = "0.13.0"
   val Enumeratum              = "1.6.1"
-  val GeoTrellis              = "3.6.0"
+  val GeoTrellis              = Def.setting(ver("3.6.0", "3.6.1-SNAPSHOT").value)
   val Jts                     = "1.16.1"
   val Monocle                 = "2.1.0"
   val Refined                 = "0.9.24"
@@ -16,23 +25,9 @@ object Versions {
   val Scapegoat               = "1.4.8"
   val Shapeless               = "2.3.7"
   val SpdxChecker             = "1.0.0"
-  val Sttp                    = "3.2.3"
-  val SttpModel               = "1.4.4"
-  val SttpShared              = "1.2.2"
+  val Sttp                    = "3.3.4"
+  val SttpModel               = "1.4.7"
+  val SttpShared              = "1.2.5"
+  val Fs2                     = "2.5.6"
   val ThreeTenExtra           = "1.6.0"
-}
-
-object Dependencies {
-
-  private def ver(for212: String, for213: String) = Def.setting {
-    CrossVersion.partialVersion(scalaVersion.value) match {
-      case Some((2, 12)) => for212
-      case Some((2, 13)) => for213
-      case _             => sys.error("not good")
-    }
-  }
-
-  def geotrellis(module: String) = Def.setting {
-    "org.locationtech.geotrellis" %% s"geotrellis-$module" % ver(Versions.GeoTrellis, "3.6.1-SNAPSHOT").value
-  }
 }